backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
firebase-admin==6.1.0
google-cloud-firestore==2.13.1
google-cloud-storage==2.10.0

# AI/ML
openai==1.3.0
langchain==0.0.340
tiktoken==0.5.1

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
httpx==0.25.1
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Monitoring
prometheus-client==0.19.0
